change determinate to long to allow larger computations

+++ JHU/615.202.81.FA15_data_structures/module_08/src/Matrix.java
@@ -16,16 +16,16 @@ public class Matrix{
     this(1);
   }
 
-  public static int Determinate(Matrix mat){
+  public static long Determinate(Matrix mat){
     int[][] subData = mat.getData();
     if (subData.length == 1){
       return subData[0][0];
     } else {
-      int sum = 0;
+      long sum = 0;
       int y = 0;
       for (int x=0; x<subData.length; x++){
         Matrix minorMatrix = mat.Minor(x, y);
-        sum = sum + ((int) Math.pow(-1, x+y) * subData[x][y] * Determinate(minorMatrix));
+        sum = sum + ((long) Math.pow(-1, x+y) * (long) subData[x][y] * Determinate(minorMatrix));
       }
       return sum;
     }
